refactor: rename service -> workerapi
