+++ redbench/internal/workerapi/config_test.go
@@ -1,4 +1,4 @@
-package service
+package workerapi
 
 import (
 	"encoding/json"
