+++ redbench/internal/workerapi/notifier_test.go
@@ -1,4 +1,4 @@
-package service
+package workerapi
 
 import (
 	"context"
