+++ redbench/internal/workerapi/server.go
@@ -1,4 +1,4 @@
-package service
+package workerapi
 
 import (
 	"context"
