+++ redbench/internal/workerapi/state_test.go
@@ -1,4 +1,4 @@
-package service
+package workerapi
 
 import (
 	"testing"
